+++ client_bonus.c
@@ -6,7 +6,7 @@
 /*   By: acherraq <acherraq@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/09 18:23:09 by acherraq          #+#    #+#             */
-/*   Updated: 2024/03/31 21:31:21 by acherraq         ###   ########.fr       */
+/*   Updated: 2024/04/07 13:55:42 by acherraq         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -91,7 +91,7 @@ int	main(int argc, char **argv)
 	act.sa_handler = &sig_handler;
 	sigaction(SIGUSR1, &act, 0);
 	pid = ft_atoi(argv[1]);
-	if (pid > MAX_PID)
+	if ((unsigned long)pid > MAX_PID)
 		handle_errors("INVALID PID");
 	send_message(pid, argv[2]);
 	return (0);
